+++ trajopt_sco/test/solver-interface-unit.cpp
@@ -49,7 +49,7 @@ TEST_P(SolverInterface, setup_problem)
 
   AffExpr aff;
   std::cout << aff << std::endl;
-  for (int i = 0; i < 3; ++i)
+  for (size_t i = 0; i < 3; ++i)
   {
     exprInc(aff, vars[i]);
     solver->setVarBounds(vars[i], 0, 10);
@@ -66,7 +66,7 @@ TEST_P(SolverInterface, setup_problem)
   solver->optimize();
 
   DblVec soln(3);
-  for (int i = 0; i < 3; ++i)
+  for (size_t i = 0; i < 3; ++i)
   {
     soln[i] = solver->getVarValue(vars[i]);
   }
@@ -98,7 +98,7 @@ TEST_P(SolverInterface, DISABLED_ExprMult_test1)  // QuadExpr not PSD
 
   AffExpr aff1;
   std::cout << aff1 << std::endl;
-  for (int i = 0; i < 3; ++i)
+  for (size_t i = 0; i < 3; ++i)
   {
     exprInc(aff1, vars[i]);
     solver->setVarBounds(vars[i], 0, 10);
@@ -108,7 +108,7 @@ TEST_P(SolverInterface, DISABLED_ExprMult_test1)  // QuadExpr not PSD
 
   AffExpr aff2;
   std::cout << aff2 << std::endl;
-  for (int i = 3; i < 6; ++i)
+  for (size_t i = 3; i < 6; ++i)
   {
     exprInc(aff2, vars[i]);
     solver->setVarBounds(vars[i], 0, 10);
@@ -126,7 +126,7 @@ TEST_P(SolverInterface, DISABLED_ExprMult_test1)  // QuadExpr not PSD
   solver->optimize();
 
   DblVec soln(3);
-  for (int i = 0; i < 3; ++i)
+  for (size_t i = 0; i < 3; ++i)
   {
     soln[i] = solver->getVarValue(vars[i]);
   }
@@ -179,7 +179,7 @@ TEST_P(SolverInterface, ExprMult_test2)
   solver->optimize();
 
   DblVec soln(2);
-  for (int i = 0; i < 2; ++i)
+  for (size_t i = 0; i < 2; ++i)
   {
     soln[i] = solver->getVarValue(vars[i]);
     std::cout << soln[i] << std::endl;
@@ -231,7 +231,7 @@ TEST_P(SolverInterface, ExprMult_test3)
   solver->optimize();
 
   DblVec soln(2);
-  for (int i = 0; i < 2; ++i)
+  for (size_t i = 0; i < 2; ++i)
   {
     soln[i] = solver->getVarValue(vars[i]);
     std::cout << soln[i] << std::endl;
